Refactor playlists component
